chore: stablish state.  Reduction now accepts mom_axes.
Going to switch up to only wrapping negative values of ``axis`` if pass in an imaginary number.
This is a bit weird, but with the additional parameters, it is not clear when should or should not
wrap negative axis.  Plus this is kinda like what numpy does with with r_